xradio xr819: added 002-xxx-RX.patch
Slightly improved latency and performance of RX packets

Dymanically decrease the sleep time when waiting for interrupts
from the radio, to reduce latency during active communication.
This helps, because we are missing some interrupts.

Print an error when xradio,xr819 is not found in DTS, which
results in running without any interrupts.

Reduce the number of xradio_bh_rx_availlen() calls, because when
called too often we are getting an error:
sunxi-mmc 1c10000.mmc: data error, sending stop command
Caused by a timeout while communicating with the radio, to query
how much data is available to read if any. The timeout is likely
a hardware issue with Orange Pi Zero and other similar boards.

Signed-off-by: Georgi Valkov <gvalkov@abv.bg>
Signed-off-by: Semen Kreyda <aswerklon@gmail.com>